src/main/java/com/hansbug/twatcher/utils/GenericPair.java
package com.hansbug.twatcher.utils;

import java.util.Objects;

public class GenericPair<T1, T2> {
    private final T1 first;
    private final T2 second;

    public GenericPair(T1 first, T2 second) {
        this.first = first;
        this.second = second;
    }

    public T1 getFirst() {
        return first;
    }

    public T2 getSecond() {
        return second;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        GenericPair<?, ?> that = (GenericPair<?, ?>) o;
        return Objects.equals(first, that.first) && Objects.equals(second, that.second);
    }

    @Override
    public int hashCode() {
        return Objects.hash(first, second);
    }

    @Override
    public String toString() {
        return "GenericPair{" +
                "first=" + first +
                ", second=" + second +
                '}';
    }
}
